Use xmlid for referencing categories
The admin user has been changed to French... so it doesn't find the
Expenses category.

Please always use xmlids and not an implicit name_get!

Also, I don't think changing the 'admin' user to be in French is
something we want to do. If files are meant to be imported in French,
just change the language in the context.

Finally, I corrected the data files to load the product templates in
every modes, not only the 'full' mode.